+++ tools/Libplanet.Explorer/Queries/ExplorerQuery.cs
@@ -5,7 +5,6 @@ using GraphQL.Types;
 using Libplanet.Blockchain;
 using Libplanet.Crypto;
 using Libplanet.Explorer.GraphTypes;
-using Libplanet.Explorer.Indexing;
 using Libplanet.Explorer.Interfaces;
 using Libplanet.Store;
 using Libplanet.Types.Blocks;
@@ -38,64 +37,20 @@ namespace Libplanet.Explorer.Queries
 
         private static IStore Store => ChainContext.Store;
 
-        private static IBlockChainIndex Index => ChainContext.Index;
-
         internal static IEnumerable<Block> ListBlocks(
             bool desc,
             long offset,
-            long? limit,
-            bool excludeEmptyTxs,
-            Address? miner)
+            long? limit)
         {
             Block tip = Chain.Tip;
             long tipIndex = tip.Index;
-            IStore store = ChainContext.Store;
 
-            if (desc)
-            {
-                if (offset < 0)
-                {
-                    offset = tipIndex + offset + 1;
-                }
-                else
-                {
-                    offset = tipIndex - offset + 1 - (limit ?? 0);
-                }
-            }
-            else
-            {
-                if (offset < 0)
-                {
-                    offset = tipIndex + offset + 1;
-                }
-            }
-
-            var indexList = store.IterateIndexes(
-                    Chain.Id,
-                    (int)offset,
-                    limit == null ? null : (int)limit)
-                .Select((value, i) => new { i, value });
-
-            if (desc)
-            {
-                indexList = indexList.Reverse();
-            }
-
-            foreach (var index in indexList)
-            {
-                Block block = store.GetBlock(index.value)
-                    ?? throw new InvalidOperationException(
-                        $"Could not find block with block hash {index.value} in store.");
-
-                bool isMinerValid = miner is null || miner == block.Miner;
-                bool isTxValid = !excludeEmptyTxs || block.Transactions.Any();
-                if (!isMinerValid || !isTxValid)
-                {
-                    continue;
-                }
-
-                yield return block;
-            }
+            var blocks = ListBlocks(
+                Chain,
+                desc ? tipIndex - offset - (limit ?? 100) : offset,
+                limit ?? 100);
+            return desc ? blocks.OrderByDescending(x => x.Index)
+                : blocks.OrderBy(x => x.Index);
         }
 
         internal static IEnumerable<Transaction> ListTransactions(
@@ -115,7 +70,7 @@ namespace Libplanet.Explorer.Queries
             }
 
             Block? block = Chain[desc ? tipIndex - offset : offset];
-            while (!(block is null) && (limit is null || limit > 0))
+            while (block is not null && limit is null or > 0)
             {
                 foreach (var tx in desc ? block.Transactions.Reverse() : block.Transactions)
                 {
@@ -210,6 +165,21 @@ namespace Libplanet.Explorer.Queries
             }
 
             return null;
+        }
+
+        private static IEnumerable<Block> ListBlocks(BlockChain chain, long from, long limit)
+        {
+            if (chain.Tip.Index < from)
+            {
+                return new List<Block>();
+            }
+
+            var count = (int)Math.Min(limit, chain.Tip.Index - from + 1);
+            var blocks = Enumerable.Range(0, count)
+                .Select(offset => chain[from + offset])
+                .OrderBy(block => block.Index);
+
+            return blocks;
         }
 
         private static bool IsValidTransaction(